requirements.txt
yfinance>=0.2.0
python-dotenv
pandas
pytz
pandas_market_calendars
flask
gunicorn
matplotlib
